#14 Move all modules to core subpackage
