+++ dataspecs/core/__init__.py
@@ -1,0 +1,7 @@
+__all__ = ["api", "specs", "typing"]
+
+
+# submodules
+from . import api
+from . import specs
+from . import typing
